Issue #145 chore: move datasource
