+++ app/src/main/java/com/gdsc/sogongsogong/data/remote/post/PostRemoteDataSourceImpl.kt
@@ -1,6 +1,7 @@
 package com.gdsc.sogongsogong.data.remote.post
 
 import com.gdsc.sogongsogong.data.api.post.PostService
+import com.gdsc.sogongsogong.data.datasource.PostDataSource
 import com.gdsc.sogongsogong.data.entity.Post
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
